chore: move services/pkg
